+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ CHANGELOG
 Unreleased
 ==========
 
+
+0.3.0
+=====
+
 * Added: Support for include_aggregate parameter for ScansApi.import_scan.
 * Added: Support for scans host-details API.
 * Changed: Model ScanDetails.hosts is now a list of ScanHost's instead of dict's.